chore(tests): move module tests from variables.rs to variables_test.rs
