+++ src/common/config/variables_test.rs
@@ -1,3 +1,5 @@
+use serde_json::json;
+
 use super::variables::parse_variables;
 
 #[test]
@@ -149,4 +151,67 @@ fn test_parse_mixed_single_and_array_values() {
     assert_eq!(arr[0].as_str(), Some("dev"));
     assert_eq!(arr[1].as_str(), Some("staging"));
     assert_eq!(arr[2].as_str(), Some("prod"));
+}
+
+#[test]
+fn plain_single_value_is_string() {
+    let vars = vec!["key=alpha".to_string()];
+    let map = parse_variables(&vars).unwrap();
+    assert_eq!(map["key"], json!("alpha"));
+}
+
+#[test]
+fn plain_repeated_key_is_array() {
+    let vars = vec!["key=alpha".to_string(), "key=beta".to_string()];
+    let map = parse_variables(&vars).unwrap();
+    assert_eq!(map["key"], json!(["alpha", "beta"]));
+}
+
+#[test]
+fn bracket_single_value_is_array_of_one() {
+    let vars = vec!["key[]=alpha".to_string()];
+    let map = parse_variables(&vars).unwrap();
+    assert_eq!(map["key"], json!(["alpha"]));
+}
+
+#[test]
+fn bracket_csv_splits_into_array() {
+    let vars = vec!["key[]=alpha,beta,gamma,delta".to_string()];
+    let map = parse_variables(&vars).unwrap();
+    assert_eq!(map["key"], json!(["alpha", "beta", "gamma", "delta"]));
+}
+
+#[test]
+fn bracket_csv_repeated_flags_combine() {
+    let vars = vec!["key[]=alpha,beta".to_string(), "key[]=gamma".to_string()];
+    let map = parse_variables(&vars).unwrap();
+    assert_eq!(map["key"], json!(["alpha", "beta", "gamma"]));
+}
+
+#[test]
+fn bracket_csv_escaped_comma() {
+    let vars = vec![r"key[]=foo\,bar,baz".to_string()];
+    let map = parse_variables(&vars).unwrap();
+    assert_eq!(map["key"], json!(["foo,bar", "baz"]));
+}
+
+#[test]
+fn mixing_plain_and_bracket_combines() {
+    let vars = vec!["key=alpha".to_string(), "key[]=beta,gamma".to_string()];
+    let map = parse_variables(&vars).unwrap();
+    assert_eq!(map["key"], json!(["alpha", "beta", "gamma"]));
+}
+
+#[test]
+fn mixing_bracket_then_plain_combines() {
+    let vars = vec!["key[]=alpha,beta".to_string(), "key=gamma".to_string()];
+    let map = parse_variables(&vars).unwrap();
+    assert_eq!(map["key"], json!(["alpha", "beta", "gamma"]));
+}
+
+#[test]
+fn plain_value_with_comma_stays_literal() {
+    let vars = vec!["key=foo,bar".to_string()];
+    let map = parse_variables(&vars).unwrap();
+    assert_eq!(map["key"], json!("foo,bar"));
 }
